# docs/swagger.yaml
basePath: /auth
definitions:
  entity.Login:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  entity.LoginResponse:
    properties:
      message:
        type: string
      token:
        type: string
    type: object
  entity.Signup:
    properties:
      birthDate:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
host: 0.0.0.0:8000
info:
  contact: {}
  description: This is Example API in Go.
  title: Go Boilerplate API
  version: "1.0"
paths:
  /health-check:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Return status of service
      tags:
      - General
  /user/v1:
    delete:
      consumes:
      - application/json
      description: Endpoint to delete user
      parameters:
      - description: Client identifier
        in: header
        name: X-Client-Id
        required: true
        type: string
      - description: Auth Token
        in: header
        name: X-Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Endpoint to delete user
      tags:
      - Delete
    put:
      consumes:
      - application/json
      description: Endpoint to update user
      parameters:
      - description: Client identifier
        in: header
        name: X-Client-Id
        required: true
        type: string
      - description: Auth Token
        in: header
        name: X-Authorization
        required: true
        type: string
      - description: Payload
        in: body
        name: Payload
        required: true
        schema:
          $ref: '#/definitions/entity.Signup'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Endpoint to update user
      tags:
      - Update
  /user/v1/login:
    post:
      consumes:
      - application/json
      description: Endpoint to login user
      parameters:
      - description: Client identifier
        in: header
        name: X-Client-Id
        required: true
        type: string
      - description: Payload
        in: body
        name: Payload
        required: true
        schema:
          $ref: '#/definitions/entity.Login'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.LoginResponse'
      summary: Endpoint to login user
      tags:
      - Login
  /user/v1/register:
    post:
      consumes:
      - application/json
      description: Endpoint to register user
      parameters:
      - description: Client identifier
        in: header
        name: X-Client-Id
        required: true
        type: string
      - description: Payload
        in: body
        name: Payload
        required: true
        schema:
          $ref: '#/definitions/entity.Signup'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
      summary: Endpoint to signup user
      tags:
      - Register
schemes:
- http
swagger: "2.0"